ws/requirements.txt
anyio==3.6.1
asyncpg==0.26.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
databases==0.6.1
fastapi==0.85.0
future==0.18.2
greenlet==1.1.3
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
lzstring==1.0.4
databases[postgresql]==0.6.1
pycryptodome==3.15.0
pydantic==1.10.2
PyJWT==2.5.0
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
psycopg2-binary
pytz
aiomysql==0.0.21
pymysql